fix(autocmd): update on InsertLeave
Making changes in insert mode does not trigger the diagnostics to
update. We don't want to do on `TextChangedI` because that would cause
it to update on every keystroke. It also is not good to use
`TextChangedI` because it gets rid of all the diagnostics as you type,
which is very disorienting. Updating on `InsertLeave` fixes the issue at
the cost of not updating until you exit insert mode, which is probably
what most users want anyway.